Converted size_t to uint32_t

+++ test/VovkPLCRuntimeWasmTestKit/simulator.cpp
@@ -138,11 +138,11 @@ WASM_EXPORT void doSomething() {
 
 template <typename T> struct OutputBuffer {
     T* data = nullptr;
-    size_t type_size = sizeof(T);
-    size_t length = 0;
-    size_t max_size = 0;
-    OutputBuffer(size_t size = 0) { init(size); }
-    void init(size_t size = 0) {
+    uint32_t type_size = sizeof(T);
+    uint32_t length = 0;
+    uint32_t max_size = 0;
+    OutputBuffer(uint32_t size = 0) { init(size); }
+    void init(uint32_t size = 0) {
         if (!size) return;
         if (data == nullptr) {
             data = new T[size];
@@ -164,7 +164,7 @@ template <typename T> struct OutputBuffer {
     T shift() {
         if (length > 0) {
             T value = data[0];
-            for (size_t i = 1; i < length; i++) {
+            for (uint32_t i = 1; i < length; i++) {
                 data[i - 1] = data[i];
             }
             length--;
@@ -173,13 +173,13 @@ template <typename T> struct OutputBuffer {
         return 0;
     }
     // Array operator getter
-    T& operator[](size_t index) {
+    T& operator[](uint32_t index) {
         if (index < length)
             return data[index];
         return data[0];
     }
     // Array operator setter (avoid overloaded 'operator[]' cannot have more than one parameter before C++2b)
-    void set(size_t index, T value) {
+    void set(uint32_t index, T value) {
         if (index < length)
             data[index] = value;
     }
@@ -196,7 +196,7 @@ WASM_EXPORT void* export_output() {
     myBuffer.push(export_request_count);
     //Print buffer contents
     Serial.print(F("Buffer <"));
-    for (size_t i = 0; i < myBuffer.length; i++) {
+    for (uint32_t i = 0; i < myBuffer.length; i++) {
         Serial.print(myBuffer[i], HEX);
         if (i < myBuffer.length - 1) Serial.print(F(", "));
     }
